avanzo con segmentacion de funcionalidades
