+++ funcion/donar.py
@@ -5,10 +5,10 @@ import sys
 import re
 from config import *
 from sisop.varios import *
-from pedirInfo import *
-from getJson import *
-from getVarios import *
-from varios import *
+from helpers.pedirInfo import *
+from helpers.getJson import *
+from helpers.varios import *
+from helpers.recursos import *
 
 def donar(s):
 	bienes = {'1': 'Viñedo', '2': 'Cantera', '3': 'Mina de cristal', '4': 'Mina de azufre'}
