+++ helpers/naval.py
@@ -1,0 +1,12 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+import re
+
+def getBarcosDisponibles(s):
+	html = s.get()
+	return int(re.search(r'GlobalMenu_freeTransporters">(\d+)<', html).group(1))
+
+def getBarcosTotales(s):
+	html = s.get()
+	return int(re.search(r'maxTransporters">(\d+)<', html).group(1))
