+++ helpers/planearViajes.py
@@ -3,7 +3,7 @@
 
 import re
 import time
-from  getVarios import *
+from helpers.getJson import *
 
 def enviarBienes(s, idCiudadOrigen, idCiudadDestino, idIsla, md, vn, mr, cr, az, barcos):
 	s.post(payloadPost={'action': 'header', 'function': 'changeCurrentCity', 'actionRequest': s.token(), 'cityId': idCiudadOrigen, 'ajax': '1'})
